+++ src/mpd_worker/mpd_worker_cache.c
@@ -82,8 +82,8 @@ bool mpd_worker_cache_init(struct t_mpd_worker_state *mpd_worker_state) {
 //private functions
 static bool _cache_init(struct t_mpd_worker_state *mpd_worker_state, rax *album_cache, rax *sticker_cache) {
     MYMPD_LOG_INFO("Creating caches");
-    long start = 0;
-    long end = start + MPD_RESULTS_MAX;
+    unsigned start = 0;
+    unsigned end = start + MPD_RESULTS_MAX;
     long i = 0;
     long album_count = 0;
     long song_count = 0;
