+++ src/mympd_api/mympd_api_albumart.c
@@ -25,15 +25,15 @@ sds mympd_api_albumart_getcover(struct t_mympd_state *mympd_state, sds buffer, s
     if (mympd_state->mpd_state->feat_mpd_albumart == true) {
         MYMPD_LOG_DEBUG("Try mpd command albumart for \"%s\"", uri);
         while ((recv_len = mpd_run_albumart(mympd_state->mpd_state->conn, uri, offset, binary_buffer, mympd_state->mpd_state->mpd_binarylimit)) > 0) {
-            *binary = sdscatlen(*binary, binary_buffer, recv_len);
+            *binary = sdscatlen(*binary, binary_buffer, (size_t)recv_len);
             if (sdslen(*binary) > MPD_BINARY_SIZE_MAX) {
                 MYMPD_LOG_WARN("Retrieved binary data is too large, discarding");
                 sdsclear(*binary);
                 offset = 0;
                 break;
             }
-            offset += recv_len;
-            if (recv_len < (int)mympd_state->mpd_state->mpd_binarylimit) {
+            offset += (unsigned)recv_len;
+            if ((size_t)recv_len < mympd_state->mpd_state->mpd_binarylimit) {
                 break;
             }
         }
@@ -44,15 +44,15 @@ sds mympd_api_albumart_getcover(struct t_mympd_state *mympd_state, sds buffer, s
         mpd_response_finish(mympd_state->mpd_state->conn);
         MYMPD_LOG_DEBUG("Try mpd command readpicture for \"%s\"", uri);
         while ((recv_len = mpd_run_readpicture(mympd_state->mpd_state->conn, uri, offset, binary_buffer, mympd_state->mpd_state->mpd_binarylimit)) > 0) {
-            *binary = sdscatlen(*binary, binary_buffer, recv_len);
+            *binary = sdscatlen(*binary, binary_buffer, (size_t)recv_len);
             if (sdslen(*binary) > MPD_BINARY_SIZE_MAX) {
                 MYMPD_LOG_WARN("Retrieved binary data is too large, discarding");
                 sdsclear(*binary);
                 offset = 0;
                 break;
             }
-            offset += recv_len;
-            if (recv_len < (int)mympd_state->mpd_state->mpd_binarylimit) {
+            offset += (unsigned)recv_len;
+            if ((size_t)recv_len < mympd_state->mpd_state->mpd_binarylimit) {
                 break;
             }
         }
